% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpwm-package.R
\docType{package}
\name{gpwm-package}
\alias{gpwm}
\alias{gpwm-package}
\title{gpwm: Motif analysis using the Misha package}
\description{
gpwm package is a collection of functions for analysis of PWM motifs using the misha package.
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

}
\keyword{internal}
