% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpwm.R
\name{gpwm.motif_enrich}
\alias{gpwm.motif_enrich}
\title{Calculate motif enrichment}
\usage{
gpwm.motif_enrich(fg, bg, global_quantiles = NULL, pattern = NULL,
  size = NULL, quantile_thresh = 0.99, min_n_fg = 4, min_n_bg = 5,
  ...)
}
\arguments{
\item{fg}{foreground motifs (output of gpwm.extract_all).}

\item{bg}{background motifs (output of gpwm.extract_all).}

\item{global_quantiles}{global quantiles for motifs (output of gpwm.get_global_quantiles). if NULL 0 the global quantiles would be automatically calculated.}

\item{pattern}{prefix for motif tracks (e.g. "motifs_10bp"), can be regular expression. Needed only if global_quantiles are not give.}

\item{size}{size of intervals in order to use the correct global quantiles.}

\item{quantile_thresh}{quantile of PWM energy that is considered a motif "hit"}

\item{min_n_fg}{minimal number of hits in foreground (motifs below this number would not be included in the hyper geometric test)}

\item{min_n_bg}{minimal number of hits in background (motifs below this number would not be included in the hyper geometric test)}

\item{...}{additional parameters for gpwm.global_quantiles}
}
\value{
data frame with number of hits for each motif in the foreground and background, together with p.value and q-value (FDR adjuster). p-value is computed using a hyper-geometric test.
}
\description{
Calculate motif enrichment
}
