% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpwm.R
\name{gpwm.extract_all}
\alias{gpwm.extract_all}
\title{Extract all motifs for intervals}
\usage{
gpwm.extract_all(pattern, intervals, colname_prefix = NULL,
  tidy = FALSE, parallel = TRUE, ...)
}
\description{
Extract all motifs for intervals
}
